% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_prep.R
\name{arithmetic_prep}
\alias{arithmetic_prep}
\alias{arithmetic_prep.default}
\title{Prep Plot for Arithmetic}
\usage{
arithmetic_prep(x, ...)

\method{arithmetic_prep}{default}(x, ...)
}
\arguments{
\item{x}{a plot object}

\item{...}{passed to other methods}
}
\value{
A \code{patchwork} object
}
\description{
This function is used internally in the patchwork arithmetic operators,
e.g. \code{|} or \code{/}.
The default method wraps plot objects in \code{wrap_elements()} as needed, preparing
the plot to be processed by the arithmetic operators.

Implement custom methods if you wish objects of other classes to be compatible
with the patchwork arithmetic operators.
}
\examples{
library(ggplot2)

p <- ggplot(mtcars, aes(x = mpg, y = hp)) + geom_point()

arithmetic_prep(p)
}
