% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout.R
\name{cell}
\alias{cell}
\title{Specify a plotting area in a layout}
\usage{
cell(t, l, b = t, r = l)
}
\arguments{
\item{t, b}{The top and bottom bounds of the area in the grid}

\item{l, r}{The left and right bounds of the area int the grid}
}
\value{
A \code{plot_cell} object
}
\description{
This is a small helper used to specify a single area in a rectangular grid
that should contain a plot. Objects constructed with \code{cell()} can be
concatenated together with \code{c()} in order to specify multiple areas.
}
\details{
The grid that the cells are specified in reference to enumerate rows from top
to bottom, and coloumns from left to right. This means that \code{t} and \code{l}
should always be less or equal to \code{b} and \code{r} respectively. Instead of
specifying cell placement with a combination of \code{cell()} calls, it is
possible to instead pass in a single string\preformatted{cells <- c(cell(1, 1, 2, 1),
           cell(2, 3, 3, 3))
}

is equivalent to\preformatted{cells < -"A##
          A#B
          ##B"
}

For an example of this, see the \code{\link[=plot_layout]{plot_layout()}} examples.
}
