% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_cell.R
\name{cellGrob}
\alias{cellGrob}
\title{Get a grob describing the content of an assemble_cell object}
\usage{
cellGrob(x)
}
\arguments{
\item{x}{An \code{assemble_cell} object}
}
\value{
A \code{gtable} object
}
\description{
Methods for this generic should be defined for all \code{assemble_cell} subclasses
and should return a compliant \code{gtable} object ready to be combined with
regular plot objects. In general it is best to call \code{cell_table()} on the
object and add grobs to this as \code{cell_table()} will return a compliant
\code{gtable}
}
\keyword{internal}
