% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide_area.R
\name{guide_area}
\alias{guide_area}
\title{Add an area to hold collected guides}
\usage{
guide_area()
}
\description{
Using the \code{guides} argument in \code{\link[=plot_layout]{plot_layout()}} you can collect and collapse
guides from plots. By default these guides will be put on the side like with
regular plots, but by adding a \code{guide_area()} to the plot you can tell
patchwork to place the guides in that area instead. If guides are not
collected or no guides exists to collect it behaves as a standard
\code{\link[=plot_spacer]{plot_spacer()}} instead.
}
