\name{karyotype}
\alias{karyotype}


\title{
karyotype plotting function
}

\description{
Plots each,color coded by chromosomal coordinate, chromosome against a 
background of the complete genome.
}

\usage{
karyotype(chr,start,end,int,ai,name,xlim,ylim)
}

\arguments{
  	\item{chr}{					%segs$chr
  	Chromosomes name of segs object. Final version of segs created from patchwork.Medians_n_AI().
	}
  	\item{start}{				%segs$start
	Start position of segment.
	}
	\item{end}{					%segs$end
  	End position of segment.	
	}
	\item{int}{					%segs$mean
  	Median.
	}
	\item{ai}{					%segs$ai
  	Allelic imbalance.
	}
	\item{name}{
  	Default is the name of the input BamFile from patchwork.plot().
  	Plot generated as "name.karyotype.png" in your working directory.
	}
  	\item{xlim}{
  	Default is c(0,2.4). Limit of x axis.
	}
  	\item{ylim}{
  	Default is c(0.3,1). Limit of y axis.
	}
}

\details{
Vertical axis: Allelic imbalance \cr
Horizontal axis: Total intensity \cr

The plot is a overview, for a closer look see the plots generated by karyotype_chroms().
}

%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}

%\references{
%% ~put references to the literature/web site here ~
%}

\author{
Markus Mayrhofer, \email{markus.mayrhofer@medsci.uu.se}
Sebastian DiLorenzo, \email{sebastian.dilorenzo@medsci.uu.se}
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{patchwork.plot}} \cr
\code{\link{karyotype_chroms}}
}

%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%Run examples of karyotype stuff. or not, maybe just point back to patchwork.plot.
%}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line