% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patchGrob}
\alias{patchGrob}
\title{Get a grob describing the content of a patch object}
\usage{
patchGrob(x, guides = "auto")
}
\arguments{
\item{x}{An \code{patch} object}
}
\value{
A \code{gtable} object
}
\description{
Methods for this generic should be defined for all \code{patch} subclasses
and should return a compliant \code{gtable} object ready to be combined with
regular plot objects. In general it is best to call \code{patch_table()} on the
object and add grobs to this as \code{patch_table()} will return a compliant
\code{gtable}
}
\keyword{internal}
