% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patchMorphSummary.R
\name{patchMorphSummary}
\alias{patchMorphSummary}
\title{patchMorphSummary}
\usage{
patchMorphSummary(pm_list, suit = -1, gap = -1, spur = -1)
}
\arguments{
\item{pm_list}{A list of patchMorph RasterLayers (pm_list).}

\item{suit}{A integer. The values of suitability over which to provide summaries. If left blank, all
values in the pm_laist object will be used.}

\item{gap}{A integer. The gap values over which to provide summaries. If left blank, all
values in the pm_laist object will be used.}

\item{spur}{A integer. The spur valuesover which to provide summaries. If left blank, all
values in the pm_laist object will be used.}
}
\value{
A rasterLayer. Sums the members of a patchMorph list (pm_list) from pm_hierarchy
}
\description{
patchMorphSummary
}
\examples{
myFile <- system.file("extdata", "mixedconifer.tif", package="patchwoRk")
myRas <- raster(myFile)

pm.layered.result.map <- patchMorph(data_in = myRas, suitVals = c(0, 1, 2),
gapVals = c(2, 6, 3), spurVals = c(2, 6, 3))
pm.layered.sum. <- patchMorphSummary(pm.layered.result)

plot(myRas, main="Original Raster")
plot(pm.layered.sum, main="PatchMorph Multi Results (Gap-2:10 & Spur-2:10)")
}
\references{
Girvetz EH, and Greco SE. 2007. How to define a patch: a spatial model for hierarchically
delineating organism-specific habitat patches. Landscape Ecology 22: 1131-1142.
}
