% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assemble.R
\name{patchworkGrob}
\alias{patchworkGrob}
\title{Convert a patchwork assemble to a gtable}
\usage{
patchworkGrob(x)
}
\arguments{
\item{x}{A \code{ggassemble} object}
}
\value{
A \code{gtable} object
}
\description{
This function is the patchwork analogue of \code{\link[ggplot2:ggplotGrob]{ggplot2::ggplotGrob()}} in that it
takes an unevaluated patchwork plot object (a \code{ggassemble}) and fixate it
into a gtable object to further manipulate directly.
}
\keyword{internal}
