% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_annotation.R
\name{plot_annotation}
\alias{plot_annotation}
\title{Annotate the final plot assemble}
\usage{
plot_annotation(title = NULL, subtitle = NULL, caption = NULL,
  tag_levels = NULL, tag_prefix = NULL, tag_suffix = NULL,
  tag_sep = NULL, theme = NULL)
}
\description{
The result of this function can be added to an assemble using \code{+} in the same
way as \code{\link[=plot_layout]{plot_layout()}}, but unlike \code{\link[=plot_layout]{plot_layout()}} it will only have an
effect on the top level plot. As the name suggests it controls different
aspects of the annotation of the final plot, such as
}
