% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme.R
\name{plot_theme}
\alias{plot_theme}
\title{Modify global aesthetics}
\usage{
plot_theme(background = NULL, border_color = "transparent",
  border_thickness = NULL, border_linetype = NULL,
  padding = unit(c(0, 0, 0, 0), "cm"))
}
\arguments{
\item{background}{The background color of everything, including everything around the plots. (default: NULL -> white or transparent)}

\item{border_color}{The color of a border around the plot (default: "transparent")}

\item{border_thickness}{The thickness of the border (default: NULL -> 0)}

\item{border_linetype}{The line type of the border (default: NULL -> solid)}

\item{padding}{Padding around the outer edge of the image. The order of the vaues is top, right, bottom, and left. (default: \code{unit(c(0,0,0,0), 'cm')})}
}
\value{
A \code{plot_layout} object to be added to a \code{ggassmble} object
}
\description{
Modify global aesthetics
}
\examples{
library(ggplot2)

p1 <- ggplot(mtcars) + geom_point(aes(mpg, disp)) + theme_bw()
p2 <- ggplot(mtcars) + geom_boxplot(aes(gear, disp, group = gear)) + theme_bw()

p1 + p2 + plot_theme(background = "gray92")
}
