% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patchwork.R
\name{patchworkGrob}
\alias{patchworkGrob}
\title{Convert a patchwork to a gtable}
\usage{
patchworkGrob(x)
}
\arguments{
\item{x}{A \code{patchwork} object}
}
\value{
A \code{gtable} object
}
\description{
This function is the patchwork analogue of \code{\link[ggplot2:ggplotGrob]{ggplot2::ggplotGrob()}} in that it
takes an unevaluated patchwork object and fixate it into a gtable object to
further manipulate directly.
}
\keyword{internal}
